create /client folder for browser-rendered app
